[fsimx6/ul/sx] U-Boot may overwrite device tree settings
In the past, when U-Boot writes some run-time device tree values on
fsimx6, fsimx6ul or fsimx6sx, it did not replace existing entries.
This was meant to give the customer an option to override the run-time
settings in the device tree. However for most of the values that we
are trying to set in this stage, this does not make sense. For example
the user can not know which ECC strength is required for a board or
what NBoot version is actually present. We are setting these values
in U-Boot exactly because we can not know this information at the
time device tree creation.

However on some values it actually makes sense that the user can
override it via the device tree. For example if he likes to rename
the board to some own name, we should not overwrite this in U-Boot.
This is for example true for values "board_name" and "platform".

Therefore add an argument "force" to each FDT function. If force is
set, the new value will overwrite any previous device tree settings,
showing "Overwriting property ... from device tree". If force is not
set, then the device tree value takes precedence if set, showing
"Keeping property ... from the device tree".