Add possibility to add function board_preboot_os()
Similar to arch_preboot_os() that gives a CPU architecture the option
for a call-back function before the OS is started, also add an optional
board_preboot_os() function, that allows to run board-specific code
before the OS is started.

By default, an empty weak function is called, which means if needed,
the board specific code can implement a regular function that is called
instead.