Fix setting of platform and sercon variables in board_late_init()
On all F&S boards there are environment variables sercon and platform.
They are set to "undef" in the default environment. When booting the
board, this content is checked in board_late_init() and replaced by
the real serial line and real platform name of the board we are currently
running on.

However the code doing this was too simple, for example

 if (strcmp(getenv("sercon"), "undef") == 0) ...

This code would fail if the variable was deleted. Then getenv() returns
NULL and strcmp() would dereference the NULL pointer, causing an
exception.

Improve the code to not fail in case of a non-existent variable by
checking for NULL.