Add support for CONFIG_NETCONSOLE to fsimx6/sx/ul/fsvybrid
Fix the code in drivers/net/netconsole.c to work with the struct stdio_dev
that is slightly modified by F&S.

Then add CONFIG_NETCONSOLE to fsimx6.h, fsimx6sx.h, fsimxul.h and fsimx6ul.h.

This means that stdin, stdout and stderr can now be redirected to a PC via
ethernet. On the PC you need to run the following tool:

  tools/netconsole <client-ip-address>

You have to give the IP address of the board. On the board, by default the
serverip is used. If you want to redirect the console to a different host,
set variable ncip to the IP address of the host PC.

  setenv ncip <host-ip-address>

Then you can start redirection:

  setenv stdout nc; setenv stdin nc

From now on all input/output is seen on the PC. To route back to the board,
use:

  setenv stdout serial; setenv stdin serial

To stop the netconsole tool on the PC, type Ctl-T.