Patch shared memory addresses for RPMsg in device tree on fsimx6sx
The Linux RPMsg driver needs a 64KB memory region shared with Cortex-M4 to
pass buffer information, the so-called rpmsg vring data. This memory region
must not be touched by Linux itself. Therefore we reduce the memory size that
is shown to Linux by 64KB. In case of 2GB RAM, it is even slightly more
complicated, because Cortex-M4 can not access RAM beyond 1.5GB. So we create
a gap of 64KB at the end of the first GB of RAM in this case.

This excluded 64KB RAM space can then be used as shared memory. However the
base address of this region varies with the RAM size that is actually
available on the board we are currently running on. So we have to inform
the kernel about the correct address. In fact the shared memory region is
split in two parts, one half for each direction of the core-to-core
communication. So we actually need to inform the kernel about two addresses.

The RPMsg driver in Linux is already modified by F&S so that these addresses
are not hard-coded anymore, but can be given in the device tree. So here we
add code to ft_board_setup() to patch these two addresses in the device tree
to reflect the currently valid values.