Set WLAN MAC address in bdinfo for fsimx6sx/fsimx6ul
If the WLAN MAC address is only an additional MAC entry in bdinfo, finding
it in Linux (for example by a script) is rather difficult. It can be MAC0,
MAC1 or MAC2, depending on the number of ETH ports that the board has. Change
the entry to "MAC_WLAN". No number is needed as it is rather unlikely that
there is ever more than one WLAN device on a board.

For boards based on the Silex QCA9377 WLAN chip, the driver also needs the
MAC address in a file of the following form, where the address is given
without colons.

  Intf0MacAddress=aabbccddeeff

Also create this entry in bdinfo in these cases (i.e. efusA9X and efusA7UL
on board revisions starting with 1.20).

While we are at it, fix the device tree path name to ETH_B on fsimx6sx. With
the wrong path, the second ethernet device was not disabled on boards where
it was not equipped.