Add option to switch off EEE for Atheros PHY
Using power saving function Energy Efficient Ethernet (EEE) may cause losing
packages for ping and similar transfers. If option CONFIG_PHY_ATHEROS_NO_EEE
is set, the PHY is instructed to not advertise the EEE feature, which will
result in links that do not use EEE. This will use more energy, but is more
stable in the tranfers.

Switching off EEE advertisement is done in register MMD7 on offset 0x3c in
bits 1 and 2.

Add option CONFIG_PHY_ATHEROS_NO_EEE to fsimx6.h and fsimx6sx.h, but leave it
commented out by default.

On CONT1, the external PHYs must be configured manually. This usually takes
place when the ethernet port is activated. But these PHYs are connected to
an external 5 port ethernet switch and are independet from the FEC ports, so
they would not be configured at all if it is not done manually.