Add CONFIG_IMX_THERMAL to fsimx6/ul/sx
CONFIG_IMX_THERMAL will read temperature range and speed grading from
fuses and shows an according output when printing the CPU version at
boot time. This is a new feature found in rel_imx_4.1.15_2.0.0_ga
that we use from now on on all boards based on i.MX6, i.MX6SX and
i.MX6UL/ULL.

Also replace Freescale with NXP in some comments and update the copyright
of the header files to 2018.