Enhance speed grade detection and temperature output on i.MX6UL/ULL
The speed detection must be modified to work on i.MX6ULL. Simplify the
evaluation in get_cpu_speed_grade_hz() by splitting the cases for i.MX6UL
and i.MX6ULL and using a switch statement in each case.

In print_cpuinfo(), improve the speed output and optimize the temperature
output.