On fsimx6ul, add platform names for variants with i.MX6ULL
In Linux we need different device trees for boards equipped with i.MX6UL and
i.MX6ULL. Similar to the i.MX6 case, where we append "dl" for Solo/DualLite
or "q" for Dual/Quad to the platform name, we now want to have the platform
name ending in 'ull' if we are on an i.MX6ULL. This means if the platform
name already ends with 'ul', we simply add a second 'l'. Otherwise we append
'ull' to the name.

  efusa7ul -> efusa7ull
  cubea7ul -> cubea7ull
  picocom1.2 -> picocom1.2ull
  cube2.0 -> cube2.0ull
  ...