Deactivate VLCD_ON and VCFL_ON in U-Boot on fsimx6/sx/ul
On some platforms, the pads for VLCD_ON and VCFL_ON are configured for
Pull-Keeper by default. This means if the signal is low at reset, it stays
low, and if the signal is high, it stays high. We should not rely on this
random behavior. So change this setting to a real pull-down to actively
disable these voltages at boot time (or pull-up in case of PicoCOMA9X, where
these signals are active-low).