Initialize DVS signal on fsimx6ul
Our DVS signal is on pad NAND_DQS. This pad is actually configured as NAND_DQS
function when U-Boot starts (maybe from CPU boot ROM because we are booting
from NAND). In this case this signal is an output with a high signal, which
enables our DVS, i.e. switches the core voltage to a lower voltage of 1.2V.
However we need DVS=0 to have full core voltage. Otherwise CPU speeds above
528MHz will not work.

So configure this pin as GPIO with pull-down to hold DVS low, i.e. core
voltage high.