On fsimx6ul, add option to limit the CPU to 528 MHz
i.MX6ULL CPUs are also available with 800 MHz and 900 MHz. To be able to use
these CPUs in applications that are tested with the regular 528 MHz version,
it might be necessary to limit the speed to 528 MHz. This can be done by
soldering an appropriate config jumper. NBoot will read the jumper and
passes the value in bit FEAT2_SPEED in chFeatures2.

U-Boot itself is only running with 400 MHz. So we only need to tell Linux
to limit the speed. The speed settings, so-called operating points, are
given in the device tree. Each point is a value pair, first value is the
frequency (in kHz) and the second value is the voltage (in mV) required for
this frequency. On i.MX6UL/ULL there are two such arrays, one array for the
ARM CPU voltage and one array for the SOC voltage.

When doing the final fixups in the device tree before starting the kernel
(see ft_board_setup() in board/F+S/fsimx6ul/fsimx6ul.c), we check these
arrays and remove all operating points that are higher than 528 MHz. So
Linux will just see points with at most 528 MHz and will not use higher
speeds.