Fix problem with slow USB sticks
The USB driver in U-Boot uses a 5s timeout for bulk transfers. If we use up
to 65534 blocks (~32 MB) in one transfer, the timeout may be triggered on
very slow USB drives. The large number is not really necessary, we can also
use a smaller number without loosing noticable transfer speed.

Reduce the number of blocks to 32768. This means we will transfer at most
16 MB in one go. This is less than 27 MBit/s, even very slow USB devices
should be capable of delivering this speed. Though it is still a little bit
faster than USB Full Speed (12 MBit/s). But mass storage devices should be
capable to use High Speed anyway.