Fix video and nulldev stdio devices
The signature for stdio devices was changed in the past, but this change
was not performed for the video device in drivers/video/cfb_console.c and
the nulldev device in common/stdio.c. Fix this now.

In addition, we must use static structures to register stdio devices, because
the data is not copied anymore. This requires some changes in drv_video_init()
and similar functions.