Replace GPL comments in F&S files with SPDX-License-Identifier
Many files written by F&S still had a long GPL license text in the comment
at the beginning of the files. Common U-Boot practice is to just use a
reference to the license text, the so-called SPDX-License-Identifier.
Remove the license text in these files and add the SPDX-License-Identifier
instead.

While at it, also remove unwanted trailing whitespace that was present in
some of these files.