Switch off backlight and display voltages when booting Linux
To avoid that the user sees display content when the display clock is off,
switch off backlight and display voltage before booting Linux. Then the
display is black for a short time until Linux restarts the display driver.
Otherwise some displays would show test patterns (for example a red screen)
when the clock is off, which is more annoying than having a short black
period.

FIXME:

This is only a temporary solution and only implemented for fsimx6. In the
end the display content should go seemlessly to Linux. But this is very
difficult to implement and will be done later.