Add fix for eMMC with activated Enhanced Mode
Usually the size of the device decides whether byte mode or sector mode is
used to address data. Devices up to 2 GB use byte mode, devices with more
than 2 GB, i.e. high capacity cards, use sector more. If the extended CSD
register is valid, it holds an entry SEC_CNT with the number of available
sectors in the user region. This entry is usually only valid if the device
is bigger than 2 GB. However if additional General Purpose partitions are
defined of if the enhanced mode is activated (switching to Pseudo Single
Level Cell mode, pSLC), then the number reported in SEC_CNT may show a value
that would indicate a device with less than 2 GB.

The old code used byte mode in this case, which is wrong. The standard
states that in this case the SEC_CNT value remains valid and sector mode
has to be used even if the size seems to be smaller than 2 GB. So using the
size alone as an indication whether to use byte mode or sector mode is not
sufficient. Instead bit 30 in the OCR shows directly what mode the card
needs. This bit is saved as mmc->high_capacity. So instead of a size
comparison, simply use this field.

In addition if enhanced mode is activated, ERASE_GRP_DEF must be set to
erase chunks of the correct size. Set this flag also in ext_csd[] array
or otherwise the bus width will not be set correcty.