F&S boards: Update loadaddr, add fdtaddr, use zImage by default
The unpacked kernel has to be loaded to offset 0x8000 from the beginning
of RAM. This is defined in the kernel. However if a packed kernel is used
(gzipped or similar), the kernel must be unpacked first. If this packed
image is already loaded to offset 0x8000, then the kernel startup-code
has to move the image away to some higher address first to be able to
unzip it to offset 0x8000. But if we directly load the kernel to a
higher address, then the kernel can skip this extra copy step. So this
is slightly faster.

Because of this we always loaded the kernel to offset 0x00800000, which
was also our default loadaddr. This was meant for images up to 8 MB and
we only had at most 5 MB. However this number has to be seen for the
*unpacked* image. And here newer, bigger kernels already exceed 8 MB.

So move the loadaddr to offset 0x01000000. This now makes room for unpacked
kernels up to nearly 16 MB. This is a compromise. We could also use
0x02000000 which would be enough even for the biggest Linux kernels. But
as this address is also our loadaddr, other images like root filesystems
will also be loaded to this offset by default. The higher the loadaddr
the less room is available in RAM behind this address and we get problems
on boards with small RAM or with large images. So just use to 16 MB. This
should nonetheless be safe for quite a while.

This also means that the loadaddr for the FDT needs to be moved, too. We
use offset 0x02000000, another 16 MB later. If the unpacked kernel can only
be 16 MB at most, then the packed image will fit into 16 MB even more.

The fdt load address was given as a constant value within several commands
in the past. Add an environment variable fdtaddr that holds this address
and use it instead. This can be changed easily by the user if he likes a
different address. Also future changes will be easier now.

As a final step change the default image format from uImage to zImage.
NXP also uses zImage in newer versions.

Remark:
We do not need to change the boot command from bootm to bootz because we
have extended bootm in the past to also understand zImage format. By
continuing to use bootm, the user can freely decide whether to use zImage
or uImage. Both will work in the same way.