Change meaning of loadaddr, always set fileaddr
Environment variable $loadaddr is a default address that can be used in
commands that need an address to load (or save) an image. F&S has made this
easy to use by accepting also "." instead of $loadaddr.

This variable should *not* be modified by the commands themeselves. Only the
user may change it. If a file is loaded successfully, variables fileaddr
and filesize are set according to the actual address and size. So if a
user calls

  tftp 13000000 uImage

then fileaddr will be set to 13000000 and filesize to the length of the file.
loadaddr will not be changed anymore, so at a subsequent

  tftp . uImage

will load the file to the original loadaddr again.

This change needs modifications in several files that used to set the
filesize and or fileaddr variables and also to the net infrastructure to
use fileaddr now and not loadaddr anymore.

- Add functions set_fileaddr(), get_fileaddr() and setenv_fileinfo()
- set_loadaddr() does not set the environment variable anymore
- Change all occurences of these functions
- Remove parse_loadaddr_base(), it is unused