USB hub handling needs some delay before resetting ports
Previous USB hub code had a 200ms delay before resetting the USB hub. The
current code of u-boot-2016.05 dropped this delay. This causes problems on
the hub on efus-SKIT, devices on the hub ports may not be detected correctly.
For example USB storage devices are only recognized as full speed instead of
high speed devices.

Re-introduce the delay again, but use 50ms only. That seems to be enough.

In function usb_stor_write(), the debug messages should show "usb_write",
not "usb_read".