Fix fsl_esdhc_cfg handling when using GPIOs for Card Detect (CD)
The regular fsl_esdhc.c code in u-boot-2016.05 changed the way how the private
data of the driver is stored and referenced via the priv entry in struct mmc.
The struct fsl_esdhc_cfg that is passed to fsl_esdhc_initialize() is copied
to the struct fsl_esdhc_priv now.

F&S used an own struct fs_mmc_cfg with additional info that also contained a
struct fsl_esdhc_cfg. We passed a pointer to the embedded struct to
fsl_esdhc_initialize() and when later called in board_mmc_getcd(), we could
get back to this information by using container_of(). Because copied data is
used now, this trick does not work anymore.

Add an entry cd_gpio to to struct fsl_esdhc_cfg that stores the GPIO number
directly. Then we do not need the surrounding struct anymore.

NXP/Freescale also handles board_mmc_getcd() incorrectly now. They still use
entry priv of struct mmc to get access to the private data and cast that to
struct fsl_esdhc_cfg. But it is struct fsl_esdhc_prov now where this pointer
points to. It is pure coincident that the code works because both structs
have the USDHC base address at the same offset.

Therefore change the code to actually use struct fsl_esdhc_cfg at the
beginning of struct fsl_esdhc_priv, so that this cast works. This also
simplifies the copying to a simple assignment of the structures, which
further increases code stability because no entries will be forgotten if
the struct changes in the future. In the previous code, the copy section
would have to be changed in such a case, too.

This modification also requires some minor changes in fs_mmc_common.[ch].