Have option to keep CONFIG_BOOTDELAY unset
F&S sets variable bootdelay to "undef" and then board startup code replaces
this value with a board specfic value. But setting undef is not possible
anymore because Kconfig expects an int here. So we are setting bootdelay as
part of the CONFIG_EXTRA_ENV_SETTINGS now and keep CONFIG_BOOTDELAY unset
to avoid a duplicate entry in the default environment. This needs an extra
boolean parent entry for CONFIG_BOOTDELAY in common/Kconfig to have the
possibility to keep CONFIG_BOOTDELAY unset. If some customer wants to set a
fix value in Kconfig, this still works.

In fsimx/sx/ul.h, fsvybrid.h, add bootdelay=undef to F&S environment settings
and remove CONFIG_BOOTDELAY and CONFIG_ZERO_BOOTDELAY_CHECK. The latter was
removed from U-Boot in the meantime and is handled by negative values for
CONFIG_BOOTDELAY now.