Change nand chip allocation in fsl_nfc_fus.c and mxs_nand_fus.c
Global array nand_info[] now only holds pointers to the MTD infos. The struct
itself is allocated as part of the nand_chip allocation. In case of F&S NAND
drivers, this is done in the driver itself. So change the above files to
work with the new concept.

In mxs_nand_fus.c, there was only one instance of ecc_layout. This could cause
problems if there is more than one nand_chip and these chips use different
ECC layouts. Therefore move this layout to struct mxs_nand_priv, which has
an instance for each NAND chip.