Remove patch_sercon() from boot code
Old Linux kernels showed some output on the serial line when uncompressing
the zImage. This code was part of the uncompressing code prepended to the
zImage, but of course was running before the real kernel started. To make
it possible to use this code on different F&S boards that use different
serial ports for console output, it was necessary to patch the serial port
to use in the zImage binary. This was done in common/bootm.c and cmd/bootz.c.

However newer kernels do not output any messages in the uncompressing code
anymore. To avoid having the non-standard macro CONFIG_SYS_PATCH_TTY, this
is the time to remove this patch code again.