Add F&S FAT driver again, available as option
The regular U-Boot FAT driver has some issues:

- It uses lots of memory space, up to two times the size of a cluster,
  which may be up to 128KB (2x 64KB).
- It does not work well with long filenames. If two files only differ in
  a character behind the 11th character (8+3), the wrong file may be loaded.

Therefore F&S has implemented an own version of the driver:

- Only one sector with 512 bytes is needed as buffer
- Wildcards are supported for ls and load
- Long filenames do always work

However at the moment there is no write support in the F&S version.

Add the F&S driver as an option. The user can decide whether to use the
original driver or the F&S version.

- Add CONFIG_FAT_ORIG and CONFIG_FAT_FUS to fs/fat/Kconfig. CONFIG_FS_FAT
  is now a generic symbol that just means "there is FAT support". The real
  driver is selected by one of the other two CONFIG settings.
- Add F&S FAT code as fs/fat/fat_fus.c and include/fat_fus.h.
- Change include/fat.h to handle both FAT drivers.
- Activate CONFIG_FAT_FUS in {fsimx6,fsimx6sx,fsimx6ul,fsvybrid}_defconfig.
- Change include/config_fallbacks.h to select the original FAT driver.
- In fs/fs.c, use generic ls for original FAT driver and the F&S ls with
  wildcards for F&S FAT driver.

Remark:

The wildcard implementation that is also part of the F&S FAT driver, was
not removed previously and is therefore not part of this commit. It
consists of include/wildcard.h and all files in directory fs/wildcard/.