memcpy.S for ARM must not modify r9
r9 is the pointer to the global data (gd). It must not be changed in any
case, even if it is restored afterwards. Otherwise if the code is interrupted
while r9 has a different value, the interrupt service routine (ISR) may use
a bad value for gd.

We had problems that tftp hung when blinking LEDs were active. The blinking
is done in ISR context. Investigation showed that the ARM variant of memcpy()
(arch/arm/lib/memcpy.S) unfortunately uses and changes r9. This was fatal.

Change the memcpy() code to use r10 instead of r9.