Clean up error message if env can not be loaded
U-Boot issues a message "Loading Environment from ..." at startup. If
loading succeeds, "OK" ia appended. But if loading fails, the load
routines for the different environment devices issue error messages by
themselves. So it is not necessary to add "Failed" in this case.
Especially because this appears after a blank line and looks like a
completely different error message.

Also shorten the warning text in env/common.c if the default environment
is loaded so that it fits on the same line. And remove the double
newline.