Skip extra check for ethaddr if set with setenv
If any of the ethaddr variables is set with setenv (or env set), this
will call the callback function on_ethadddr(). This will parse the new
value and then calls eth_write_hwaddr(). However eth_write_hwaddr()
compares the new value with the value of the environment variable and
complains, if different. This is meant for the call in eth_initialize(),
where dev->enetaddr holds the default address read from the EEPROM of
the ethernet device. And if different, then the value of the environment
variable should be used and a warning is issued.

However in our case, the comparison still uses the old value of the
environment variable. Which means the values will always be different if
set anew. But then the complaint is wrong and using the old value is
also wrong. We call this function to set the new value, not to keep the
old value.

The solution is to move the real setting of the hardware address from
eth_write_hwaddr() to a new function _eth_write_hwaddr() and only call
this part in on_ethaddr(), i.e. skip the whole comparison part.