On GAR2 do not switch off external PHY on Ethernet switch
GAR2 has one (optional) external PHY connected to the Marvell Ethernet
switch. As we do not shut down the switch when booting Linux, we also
should not shut down this PHY. Therefore remove the PHY from the list of
active PHYs on the MII bus again after initializing.

As a side effect, the PHY will not appear in mdio list anymore. But it
can still be accessed with mii commands on address 9.