Fix parsing of mii devices in mii command
The mii command can be used repeatedly. That means the command is
repeated by just pressing Return in the command shell. For this
purpose, the mii command assumed that it needs to store the last
values so that they can be re-used in the repeated case. Therefore
parameter parsing was done at a central point at the beginning of
the function without looking at the type of sub-command first.

This had the result, that the <bus_name> argument of a

  mii device <bus_name>

command was also parsed as a PHY address. If the name did not
resolve in a hex number, for example "indirect mii", then the
"address" was zero and was accepted. However a bus name of "FEC"
was parsed as a hex number and the resulting address 4076 decimal
failed the range test 0-31. Thus this command was rejected.

In fact all the saving stuff for the repeated case is completely
unnecessary. In such a case, the shell passes the previous command
line to the mii command again via argv[]/argc, and the mii command
can parse the arguments again.

So remove all saving stuff, re-organize the parsing code and make
output of mii read more similar to mdio read. Especially use
0x-prefix for shown hex numbers where possible.