Better argument parsing and error handling in mdio.c
The mdio command can be called repeatedly by just pressing Return in
the command shell. Therefore the mdio command assumed that it needs to
store all values so that they can be re-used in the repeated case. But
this is not necessary because the shell will pass the previous command
line again in this case, i.e. the mdio command can parse all arguments
again.

So remove all saving stuff. This also makes the argument parsing easier
and we can get rid of function extract_phy_range() completely. This also
avoids a second evaluation of the Ethernet device in case of sub-commands
rx and wx. Move function extract_range() next to function extract_reg_range(),
they belong together.

Use CMD_RET_USAGE and CMD_RET_FAILURE to distinguish cases where simply
the command syntax was wrong and cases where the command execution fails.
Also handle some more cases as errors when the argument count is wrong.