Handle self-clearing Port Reset bit in ehci-hcd
The USB controllers on NXP CPUs have a self-clearing Port Reset bit in
register PORTSC1 (port status). This differs to the EHCI standard where
this bit has to be cleared manually. When this bit clears faster than the
50 ms that the ehci-hcd driver expects, then the original code
inadvertently clears the new speed settings that are already latched in
this register at end of reset. This switches the host controller back
to Full Speed mode, but the USB device is still in High Speed mode,
resulting in a non-working communication (analyzer shows "Reset TIMEOUT").
This causes a lot of trouble with many USB sticks.

We assume that this different behaviour of the Port Reset bit is part of
the TDI implementation of the EHCI controller. So add code to handle the
self-clearing bit in this case.

In function handshake(), the delay is 5us per loop, so we also have to
decrease the remaining time by 5, not by 1.

While we are at it, delete an obsolete comment about timer tick
computation.