Port drivers/crypto/fsl from NXP UBoot 2018.03
There is a problem with the dek_blob creation for secure boot. If you
are trying to authenticate an encrypted image, it always occurs a HAB
event. The problem is the dek_blob creation. Therefore we ported it from
official NXP UBoot 2018.03 and now it works. You can find the source of
the NXP UBoot at the following address.

Source: source.codeaurora.org/external/imx/uboot-imx/
Tag: rel_imx_4.14.98_2.0.0_ga