Allow percentage sizes and hex numbers in mmc hwpartition
When creating GP hardware partitions or setting enhanced mode for the user
area of an eMMC with the mmc hwpartition command, in the past start
addresses and sizes had to be given in decimal. This is strange because
U-Boot usually takes hexadecimal numbers for this. So change the behaviour
that hex values can at least be given by prefixing them with 0x.

In addition it is rather complex to actually determine the correct sizes.
There are no functions to request the erase block size or even the full
size of the eMMC memory. In most cases the user simply wants to use a
specific percentage. By adding a % character, you can now give sizes in
percent instead of absolute values.

Examples:

1. Set the whole user area to enhanced mode:

    mmc hwpartition user enh 0 100%

2. Create two GP partitions with 10% and 20% respectively

    mmc hwpartition gp1 10% enh gp2 20% enh

3. Create a GP partition with 40% and switch the remaining user space
   to enhanced mode:

    mmc hwpartiton gp1 40% enh user enh 0 60%

Remark: Percent values only work for enhanced areas, as the final sizes
are computed relative to the maximum enhanced size.

FIXME

Because of rounding effects, the sum of the partitions may be slightly
lower than expected.

Example:

Assume the maximum enhanced size value of some eMMC is 458 erase blocks.
You want to create two GP partitions with 20% each, and the whole user
area should be switched to enhanced mode, too. You would expect this to
be 60%.

  mmc hwpartition gp1 20% enh gp2 20% enh user enh 0 60%

However 20% of 458 is 91.6, 60% of 458 is 274.8. So if we would round up
these values, you would end up with 92 + 92 + 275 = 459, which is too
much. This would cause the command to fail. So we deliberately round down
such values. Then you end up with 91 + 91 + 274 = 456. So two erase blocks
will not be converted to enhanced mode. This is not perfect but at
least the setting should succeed.