Enable MMC as bootdevice
Enable MMC functionality for fsimx6, fsimx6sx and fsimx6ul. If
CONFIG_ENV_IS_IN_MMC is defined, the function board_mmc_get_env_dev()
will read the bootdevice from fuses and store it in mmc_boot_device. The
MMC devices are named in order of their initialization (MMC0, MMC1, ..),
so the number must be ported from USDHC device to the position in
initialization. For Vybrid the MMC-functions must be excluded. The
memory layout is shown in the header files (include/configs/fsimx6.h).