Keep .kernel_ubifs and .fdt_ubifs for Cubes
Do not switch all boards to the new macros ROOTFS, KERNEL and FDT. These
macros were introduced to have an easy option to switch between NAND and
eMMC based boards. However the Cubes have a different setting where FDT
and KERNEL are located in the rootfs, so the standard macros would be
wrong.

We also set the rootfs manually to .rootfs_ubifs in these cases as
these boards are not capable of using eMMC and if it will ever be
changed, they need other changes, too.