Revert back to F&S version of fsl_esdhc.c
The F&S infrastructure needs all elements of the fsl_esdhc_cfg. So
do not copy any values to the fsl_esdhc_priv, just embed fsl_esdhc_cfg
structure in it. This code was overwritten when merging i.MX8 code, so
restore the previous version with all necessary changes for i.MX8.

Also fix some places in the parts for MMC_SUPPORTS_TUNING where we
forgot to change the access to the regs entry the last time.