Revert to F&S timer implementation on i.MX
The original NXP timer code uses the 32MHz clock that does not allow short
delays below 1ms. This makes functions like udelay rather useless. If
CONFIG_MXC_GPT_HCLK is set, then it uses a higher clock. But the code
needs a complicated decision to use either the IPGPER clock or the 24MHz
and to determine the correct divider for 3 MHz. This is not necessary.
The IPGPER clock speed can be queried with mxc_get_clock(), so determining
the correct divider is no problem.

In addition, the NXP code uses a variable in RAM. This is not allowed
because the timer is initialized rather early in the boot sequence, before
RAM is officially available. We *must* only use global_data here.

Finally, the NXP code uses a loop to create some delay until the software
reset is completed. This is not reliable and may not work on fast CPUs
anymore. In fact the bit will automatically be reset when the reset is
over, so we can actually poll it. This is a much better way to wait until
the reset is completed.

Therefore revert to the simpler F&S implementation again that always uses
the IPGPER clock and a timer frequency of 3 MHz.

The architecture independent part of the global_data has a timer variable
already, called timebase_h/l. No need to have a second timer variable
tbu/tbl in the ARM specific global_data. This existed in the past, but is
long gone. But it seems that NXP has missed this and still uses the old way.
Change all NXP references to use the architecture independent variable and
remove the ARM specific version again.