Improve nand environment support for iMX8MN
NXP uses a weak function called "env_get_offset" to override the define
"CONFIG_ENV_OFFSET". If NAND boot is used the new offset is 60 MB. For
us it is unclear why NXP uses 60 MB offset for NAND environment,
through this reason we decided to uncomment this function if we build
our target "FSIMX8MN". It is not the cleanest way but at the moment we
don´t know how to fix it in a correct way.