Remove stdio_dev from F&S serial devices and netconsole
The original U-Boot code from eight years ago needed the following sequence
of function calls to output a character to the serial console.

1. putc(c) in console.c calls fputc(stdout, c).
2. fputc() in console.c uses the file number stdout to look up the
   active stdio_device entry, which is the serial port. There it calls
   putc(stdio_device, c).
3. This calls the wrapper function serial_stub_putc() in serial.c. There
   the priv pointer from the stdio_dev structure is used to get to the
   according serial_device. Call its putc(c).
4. In the serial device driver, there is a wrapper function for each
   available serial port. This adds the register base address and calls
   the generic function putc(uart_regs, c).
5. The generic driver function finally sends the character to the serial
   UART port by accessing the registers referenced by the base address.

F&S has improved this chain considerably. The stdio_device was embedded
in the serial_device structure and the console driver directly called
the final driver function. By passing the stdio_device as an additional
argument, the final function could convert the stdio_device pointer to
the serial_device pointer (container_of) and could find the register
base address there. So the wrapper functions in steps 3 and 4 were
completely obsolete.

This change was also performed on the netconsole driver and the video
drivers. All of them got an additional argument for the stdio_device
that allowed them to output to the according device without the need
for any additional wrapper functions.

However in the meantime, the console code has changed considerably in
U-Boot. The previous concept is only a legacy driver system. The new
Driver Model has a more generic concept with a udevice structure. This
structure can be used directly in the serial drivers and there is no
need for a separate serial_device structure anymore. And new features
were added in the intermediate layers. For example there is an option
to add an input buffer to avoid lost characters when doing copy/paste.
So skipping these layers is not a good idea anymore.

This is why we revert some changes to get closer to the current original
code. But it is not possible, to use exactly the original code, because
the serial_mxc.c driver for i.MX CPUs uses a hard coded UART register
address now. On F&S boards however, NBoot provides the address for the
UART port to use, so we need more flexibility here. We need to
be able to use any of the available UART ports. So in the legacy case,
we still pass the serial_device to the driver functions. And in the
serial_initialize() function, we call board_serial_base() that returns
the base address for the specific UART port to use.

The changes is detail:

- In cmd/console.c, change the code to iterate over the stdio_devices
  because the item linkage is back to the original linked list.
- Revert common/console.c to the original U-Boot code.
- Revert common/stdio.c to the original U-Boot code.
- Revert drivers/net/netconsole.c to the original U-Boot code.
- Revert include/iomux.h to the original U-Boot code.
- Revert include/stdio_dev.h more or less to the original U-Boot code,
  keep some const attributes and macro DEV_NAME_SIZE.
- In include/serial.h, change struct serial_device to pass the
  serial_device as additional argument. Change signatures of
  functions serial_* accordingly. Drop to_serial_device() and
  signatures for get_serial_device() and _serial_*().
- In drivers/serial/serial.c, pass the serial_device to the driver
  function calls. Also in serial_stdio_init(), add the serial
  stdio_device right here (instead of in drv_system_init() in stdio.c).
  Move get_current() more to the front, it is needed earlier.
- In drivers/serial/serial_{mxc,vybrid}.c, use the serial_device as
  parameter, not the stdio_device. Add function default_serial_console()
  to return the appropriate serial_console. This looks up the device by
  via register address given from board_serial_base(). Only register this
  one serial port, the other ports are not configured in the iomuxc
  anyway. Adding also iomuxc information would make things too
  complicated. Hey, this is the legacy driver that will be dropped soon!
- In board/F+S/common/fs_board_common.c, replace default_serial_console()
  by board_serial_base() and just return the UART register base address
  from NBoot. Drop get_debug_port() and use the serial_device name when
  setting environment variable sercon.
- fsimx6.h, fsimx6sx.h, fsimx6ul.h, fsvybrid.h: Remove CONFIG_CONS_INDEX,
  CONFIG_SYS_UART_PORT, CONFIG_SYS_SERCON_NAME, add CONFIG_MXC_UART_BASE
  instead.