Improve DRAM size configuration for fsimx8mx
The DRAM size was hard coded in the config header file.
We now support picocoremx8mx with different DRAM sizes so we have to set
it more flexibly.

The DRAM size is now set by the spl.
PHYS_SDRAM_SIZE is still used if it is  set in config header file to be
backward compatible.
Memtest is not supported at the moment.