Also use memalign_simple() in SPL, not only malloc_simple()
In the first part of SPL, when no RAM is available, a simple version
for a malloc-pool can be used, that can only allocate, but not de-allocate.
This actually worked for calls of malloc(), but some device drivers also
call memalign(), and in this case still the regular function was called,
which failed, because the regular malloc pool needs RAM to work.

Insert a short check to the regular memalign() function so that
memalign_simple() is called if the regular malloc-pool is not available
yet. This is now similar to malloc().